Moved book components to features/books/components
